# Speed up the `map()` and `forEach()` functions in DenseMatrix.js by over 300%