… test/unit-tests/type/matrix/DenseMatrix.test.js
… test/unit-tests/type/matrix/ImmutableDenseMatrix.test.js