I optimized the `map()` and `forEach()` functions in DenseMatrix.js to make them run over 300% faster!
There are multiple optimization tricks I used in conjunction to achieve this:
 * Using an iterative approach instead of an inefficient recursive approach by using an index marching algorithm
 * Duplicating the matrix into the result matrix before applying the callback function to each of the elements
 * Not using the helper `.get()` and `.set()` functions
 * Inlining the index marcher
 * Keeping a reference to each dimension of the matrix for faster access to the matrix's elements
 * Using a simple for loop when accessing the last dimension of the matrix to reduce index marches. This also means that there is no more need to store a reference to the last dimension of the matrix.
 * Using a `Uint32Array()` for the index instead of a regular `List`

<sub>Credits and many thanks to @RandomGamingDev for coming up with the algorithms used for this optimization!<sub>