@RandomGamingDev 2024-08-14
This PR was created in response to this: https://github.com/josdejong/mathjs/pull/3149#pullrequestreview-1885544599

> I didn't find any performance issues with most functions or operators. I noticed though that the .map() method of DenseMatrix is relatively slow. This is a off topic I think, we should address that in a separate PR (if there is room for improvement). In any case, I would like to proceed with this PR.