@josdejong 2024-08-21
I've added a benchmark to test the performance of the map and forEach method, see https://github.com/josdejong/mathjs/commit/88a4b35e9d7bbd4bc2e2bdaba5c20d3da702d870. Running that I see that the `map` method is about 15 times faster which is incredible! The `forEach` method is about 1.5 times faster. This looks really promising!

A number of feedbacks:

1. I prefer keeping the API the same: passing a plain and simple `Array` as `index`. It would be a breaking change, and I think it would be a bit confusing since `Uint32Array` is not a commonly used type. So I prefer replacing the `Uint32Array` array with a regular `Array` again. I'm curious to see if that deteriorates the performance a lot again (I expect not).
2. The `index` was immutable, and in this PR it is mutated every time. This is great for performance, but can cause weird errors when storing these indexes for use afterwards. I prefer the API to be "worry free" by keeping the indices immutable. This means we need to create a new array every time we call the callback. Can you give that a try and see how much impact that has on the performance? Also: we'll need to add unit tests that verify that the index is not mutated.
3. The code in `map` and `forEach` is duplicated. Can we do something about that and reuse the logic?