@josdejong 2024-08-29
I just realize there is another PR putting effort in improving `map` and `forEach`: #3256. Let's make sure to align/unify the efforts.