@Galm007 2024-08-29
Hi @josdejong, I used the benchmark that you've added and was surprised to see that changing it back to an `Array` became faster than using a `Uint32Array`. That result combined with the benefit of keeping the API the same made it the better option.

`Array` vs `Uint32Array` -- Array is **~231%** faster in forEach() and **~101%** faster in map()
![image](https://github.com/user-attachments/assets/f501a839-c789-4b50-8523-bea11f538436)

Although in a separate benchmark done by @RandomGamingDev using p5.js, he made a simple stress test that showed Uint32Arrays significantly outperforming simple Arrays:
https://editor.p5js.org/PotatoBoy/sketches/nU-dJfVKN

Next, I changed way I cloned the index array, going from using `[...index]` to `index.slice(0)` which was significantly faster.

`[...index]` vs `index.slice(0)` -- Using `index.slice(0)` is **~110%** faster in map() and **~123%** faster in forEach()
![Screenshot_20240829_131427](https://github.com/user-attachments/assets/d8429dbd-86e2-4537-8d49-cb4124fef9cf)

Finally, I tested to see how passing a immutable index array to the callback function in `map()` would affect its performance. I found that cloning the index array deteriorated the speed by about HALF of its mutable counterpart. I only tested this in `map()` because using a mutable index in `forEach()` results in errors when running `npm test`.

Mutable index vs Cloned index using `index.slice(0)` (`map()` only) -- Using mutable index is **~161%** faster
![Screenshot_20240829_134302](https://github.com/user-attachments/assets/607ddf43-7417-4264-adfb-77ac8979ebe4)

As for dealing with the duplicated logic, I am currently experimenting with different ways of refactoring the code in ways that do not significantly deteriorate the performance. Simply calling `forEach()` inside `map()` is not feasible because it would require using the matrix's helper `.set()` function which is slow because it does not use the optimization trick of keeping track of each dimension of the matrix for faster access to each element.
![image](https://github.com/user-attachments/assets/24522c02-ecf1-46aa-a9bd-9dac0e154b02)

Performance impact of naively using `.set()` inside `map()` to resolve duplicated logic -- Naive method is **~455%** slower
![image](https://github.com/user-attachments/assets/265510a7-4ffb-4f68-81c2-ef4636e2f6d4)