@dvd101x 2024-08-29
> I just realize there is another PR putting effort in improving `map` and `forEach`: #3256. Let's make sure to align/unify the efforts.

Hi Jos and Jmar, for #3256 I intentionally left out any modifications on the `recurse` function that might affect speed due to this PR. The only significant change is that I moved the `recurse` function to the [src/utils/array.js](src/utils/array.js) method so that it would be taken when needed by other functions (`map`, `forEach`, `dense.map`, `dense.forEach`), taking into account Jos's comments on this PR.

The main speed benefit in #3256 is the reduced callback functions and the rest of the changes that I have planned is just to reduce repetition of code on the transform functions by making a utility that transforms callbacks and multiple callbacks.

For this PR my recommendation would be to also take into account speed improvements in arrays like in `map(array, callback)` and `forEach(array, callback)`. At the moment I'm not sure if it's possible to include these benefits also in the multiple callback territory like `map(array1, array2, callback(v1, v2))`.

Awesome work on the speed improvements on this PR!