@josdejong 2024-08-30
Thanks for the updates @Galm007 , this sounds promising! It makes sense that cloning of the index causes a performance hit, but I think it is necessary, and I'm glad to see it's not _that_ bad.

@dvd101x indeed JMar's PR (this one, #3251) and yours (#3256) tackle different performance bottlenecks on the same two methods, which is great. I expect some merge conflicts though, so good to be aware of the two initiatives😅.