@josdejong 2024-09-11
@Galm007 I've just merged #3256 containing performance improvements in `map`, `filter`, and `forEach`. This gives some merge conflicts. Please let me know if you need help fixing the conflicts.

There are now also a new new benchmarks in `test/benchmark/map.js` and `test/benchmark/forEach.js` that you can use to test the improvements that you're working on in this PR, that may be helpful.