@Galm007 2024-09-13
Hi @josdejong :) I've ultimately decided to simply put the duplicated logic in a function named `_forEach()` which is used by the `map()` and `forEach()` functions. Since javascript does not pass variables by reference, I made `_forEach()`'s callback return an array and index. This one indirection has negligible performance on the overall impact and fixes the issue of duplicated logic. Maybe this can also open up the possibility of replacing the helper `recurse()` function.