@Galm007 2024-09-27
Hi @josdejong, thanks for being patient with me, there were personal circumstances that popped up which I had to focus on for the moment. I resolved the feedback suggestions you gave me and it is great that the performance gains still stayed significantly high even after changing the `_forEach()` function to using immutable index arrays. Have an awesome day!