@josdejong 2024-09-27
Thanks for the updates Jmar. The delay is no problem at all, please take care of personal life first!

I ran the `/test/benchmark/map.js` test before and after this PR, the performance improvement is huge! 😎 

```text
# Before

abs(genericMatrix)                         x 51,779 ops/sec ±0.40% (97 runs sampled)
abs(array)                                 x 586,908 ops/sec ±0.51% (95 runs sampled)
abs(numberMatrix)                          x 47,169 ops/sec ±0.95% (95 runs sampled)
genericMatrix.map(abs)                     x 40,415 ops/sec ±0.69% (94 runs sampled)
numberMatrix.map(abs)                      x 38,059 ops/sec ±0.48% (94 runs sampled)
map(genericMatrix, abs)                    x 40,118 ops/sec ±0.47% (97 runs sampled)
map(numberMatrix, abs)                     x 37,813 ops/sec ±0.49% (95 runs sampled)
map(array, abs)                            x 45,788 ops/sec ±0.89% (93 runs sampled)
map(array, abs.signatures.number)          x 67,965 ops/sec ±0.48% (95 runs sampled)
genericMatrix.map(abs.signatures.number)   x 55,406 ops/sec ±0.65% (96 runs sampled)
numberMatrix.map(abs.signatures.number)    x 51,242 ops/sec ±0.47% (95 runs sampled)

# After

abs(genericMatrix)                         x 349,299 ops/sec ±0.68% (91 runs sampled)
abs(array)                                 x 651,351 ops/sec ±0.68% (96 runs sampled)
abs(numberMatrix)                          x 350,675 ops/sec ±0.64% (93 runs sampled)
genericMatrix.map(abs)                     x 153,168 ops/sec ±0.54% (95 runs sampled)
numberMatrix.map(abs)                      x 151,669 ops/sec ±0.80% (94 runs sampled)
map(genericMatrix, abs)                    x 150,948 ops/sec ±0.65% (92 runs sampled)
map(numberMatrix, abs)                     x 147,245 ops/sec ±0.52% (97 runs sampled)
map(array, abs)                            x 49,053 ops/sec ±0.68% (94 runs sampled)
map(array, abs.signatures.number)          x 71,779 ops/sec ±0.52% (94 runs sampled)
genericMatrix.map(abs.signatures.number)   x 402,618 ops/sec ±0.48% (94 runs sampled)
numberMatrix.map(abs.signatures.number)    x 400,955 ops/sec ±0.63% (90 runs sampled)
```

Merging the PR now.